Moved File fund_facilitator_account in bin
